backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
pydantic==2.5.0
sqlalchemy==2.0.23
redis==5.0.1
google-generativeai==0.8.3
beautifulsoup4==4.12.2
requests==2.31.0
celery==5.3.4
psycopg2-binary==2.9.9
python-multipart==0.0.6
httpx>=0.23.0
duckduckgo-search==5.3.1b1